pyproject.toml
[project]
name = "spotdownloader"
version = "0.1.0"
description = ""
authors = [
    {name = "Ali Mohammadi",email = "alimohammadi.8773@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "yt-dlp (>=2025.8.11,<2026.0.0)",
    "python-telegram-bot (>=22.3,<23.0)",
    "spotipy (>=2.25.1,<3.0.0)",
    "python-dotenv (>=1.1.1,<2.0.0)",
    "beautifulsoup4 (>=4.13.4,<5.0.0)"
]

[tool.poetry]
package-mode = false

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
